% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{A-class}
\alias{A-class}
\alias{A}
\title{\code{A} class}
\description{
A test class, superclass of test class \code{C}.
}
\details{
This class is part of a pair of classes for testing an issue with S4
method dispatch.
}
