% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{C-class}
\alias{C-class}
\alias{C}
\title{\code{C} class}
\description{
A test class, subclass of test class \code{A}.
}
\details{
This class is part of a pair of classes for testing an issue with S4
method dispatch.
}
