% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ExpData-class}
\alias{ExpData-class}
\alias{ExpData}
\title{\code{ExpData} class}
\description{
Virtual superclass of data classes.
}
\details{
This class union allows to store any of these in a slot
of another class as well as defining common methods for all of them.
}
\seealso{
\code{\link{matrix}},
\code{\linkS4class{dgCMatrix}},
\code{\linkS4class{ExpressionSet}},
\code{\linkS4class{SummarizedExperiment}},
\code{\linkS4class{SingleCellExperiment}},
\code{\linkS4class{SpatialExperiment}}
}
