% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\docType{data}
\name{expkg_exported_value}
\alias{expkg_exported_value}
\title{Exported docs}
\format{
An object of class \code{character} of length 1.
}
\usage{
expkg_exported_value
}
\description{
Exported docs
}
\keyword{datasets}
